distance feature query on date field
